requirements.txt
azure-ai-projects~=1.0.0b7
azure-identity
python-dotenv